split db into 3 dbs
