+++ doc/requirements_py3.5.txt
@@ -1,4 +1,4 @@
-# FL-25-Feb-2016
+# FL-07-Mar-2016
 # pip requirements file for HSNmail with a Python-3.5.1
 # setuptools, pip, and venv are already present in the (virtual) python3.5
 # python-devel and mysql-devel were needed for installing mysqlclient
@@ -6,7 +6,7 @@
 # $ pip install -r requirements_py3.5.txt
 
 future >= 0.15.2
-Django >= 1.9.2
+Django >= 1.9.4
 mysqlclient >= 1.3.7
 pycups >= 1.9.73
 django-registration >= 2.0.4
